+++ rmpv-tests/tests/decode.rs
@@ -139,7 +139,10 @@ fn pass_str_from_value() {
 
 #[test]
 fn pass_bin_from_value() {
-    assert_eq!(ByteBuf::from(&[0, 1, 2][..]), from_value(Value::from(vec![0, 1, 2])).unwrap());
+    assert_eq!(
+        ByteBuf::from(&[0, 1, 2][..]),
+        from_value::<ByteBuf>(Value::from(vec![0, 1, 2])).unwrap()
+    );
 }
 
 #[test]